added simplesInterest service unit test mock and make test target

+++ business/core/simpleInterest/svc.go
@@ -2,9 +2,19 @@ package simpleinterest
 
 import (
 	"context"
+	"errors"
+	"time"
 
 	"github.com/atyagi9006/interest-calculator-svc/business/core/entities"
 	"github.com/atyagi9006/interest-calculator-svc/business/core/simpleInterest/repo"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+)
+
+const (
+	errInvalidPrincipal  = "invalid principal"
+	errInvalidROI        = "invalid roi"
+	errInvalidTimePeriod = "invalid timeperiod"
+	errInvalidId         = "invalid input"
 )
 
 type SimpleinterestSVC struct {
@@ -18,8 +28,14 @@ func NewService(r repo.Repository) *SimpleinterestSVC {
 }
 
 func (siSVC *SimpleinterestSVC) CreateSimpleInterest(ctx context.Context, si *entities.SimpleInterest) error {
-	siSVC.calculateSI(si)
-	_, err := siSVC.repo.Create(ctx, si)
+	err := siSVC.calculateSI(si)
+	if err != nil {
+		return err
+	}
+	si.ID = primitive.NewObjectID()
+	si.CreatedAt = time.Now()
+	si.UpdatedAt = time.Now()
+	_, err = siSVC.repo.Create(ctx, si)
 	if err != nil {
 		return err
 	}
@@ -29,6 +45,10 @@ func (siSVC *SimpleinterestSVC) CreateSimpleInterest(ctx context.Context, si *en
 }
 
 func (siSVC *SimpleinterestSVC) GetSimpleInterest(ctx context.Context, id string) (*entities.SimpleInterest, error) {
+	if id == "" {
+		return nil, errors.New(errInvalidId)
+	}
+
 	val, err := siSVC.repo.Read(ctx, id)
 	if err != nil {
 		return nil, err
@@ -39,7 +59,9 @@ func (siSVC *SimpleinterestSVC) GetSimpleInterest(ctx context.Context, id string
 }
 
 func (siSVC *SimpleinterestSVC) DeleteSimpleInterest(ctx context.Context, id string) error {
-
+	if id == "" {
+		return errors.New(errInvalidId)
+	}
 	err := siSVC.repo.Delete(ctx, id)
 	if err != nil {
 		return err
@@ -48,7 +70,17 @@ func (siSVC *SimpleinterestSVC) DeleteSimpleInterest(ctx context.Context, id str
 	return nil
 
 }
-func (siSVC *SimpleinterestSVC) calculateSI(si *entities.SimpleInterest) {
+func (siSVC *SimpleinterestSVC) calculateSI(si *entities.SimpleInterest) error {
+	if si.Princpal <= 0.0 {
+		return errors.New(errInvalidPrincipal)
+	}
+	if si.ROI <= 0.0 {
+		return errors.New(errInvalidROI)
+	}
+	if si.TimePeriod <= 0.0 {
+		return errors.New(errInvalidTimePeriod)
+	}
 	si.InterestAmount = (si.Princpal * si.ROI * si.TimePeriod) / 100
 	si.FinalAmount = si.Princpal + si.InterestAmount
+	return nil
 }
